+++ capsules/src/app_flash_driver.rs
@@ -39,7 +39,7 @@ pub const DRIVER_NUM: usize = driver::NUM::AppFlash as usize;
 mod ro_allow {
     pub const BUFFER: usize = 0;
     /// The number of allow buffers the kernel stores for this grant
-    pub const COUNT: usize = 1;
+    pub const COUNT: u8 = 1;
 }
 
 #[derive(Default)]
