+++ capsules/src/hmac.rs
@@ -34,14 +34,14 @@ mod ro_allow {
     pub const DATA: usize = 1;
     pub const COMPARE: usize = 2;
     /// The number of allow buffers the kernel stores for this grant
-    pub const COUNT: usize = 3;
+    pub const COUNT: u8 = 3;
 }
 
 /// Ids for read-write allow buffers
 mod rw_allow {
     pub const DEST: usize = 2;
     /// The number of allow buffers the kernel stores for this grant
-    pub const COUNT: usize = 3;
+    pub const COUNT: u8 = 3;
 }
 
 use core::cell::Cell;
